extract DeterministicXXXTransformers into separate file
