+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/transformation/LincheckClassVisitor.kt
@@ -163,168 +163,6 @@ internal class LincheckClassVisitor(
     }
 
     /**
-     * Replaces Object.hashCode and Any.hashCode invocations with just zero.
-     * This transformer prevents non-determinism due to the native hashCode implementation,
-     * which typically returns memory address of the object. There is no guarantee that
-     * memory addresses will be the same in different runs.
-     */
-    private inner class DeterministicHashCodeTransformer(
-        fileName: String,
-        className: String,
-        methodName: String,
-        adapter: GeneratorAdapter,
-    ) : ManagedStrategyMethodVisitor(fileName, className, methodName, adapter) {
-        override fun visitMethodInsn(opcode: Int, owner: String, name: String, desc: String, itf: Boolean) =
-            adapter.run {
-                if (name == "hashCode" && desc == "()I") {
-                    invokeIfInTestingCode(
-                        original = {
-                            visitMethodInsn(opcode, owner, name, desc, itf)
-                        },
-                        code = {
-                            invokeStatic(Injections::hashCodeDeterministic)
-                        }
-                    )
-                } else if (owner == "java/lang/System" && name == "identityHashCode" && desc == "(Ljava/lang/Object;)I") {
-                    invokeIfInTestingCode(
-                        original = {
-                            visitMethodInsn(opcode, owner, name, desc, itf)
-                        },
-                        code = {
-                            invokeStatic(Injections::identityHashCodeDeterministic)
-                        }
-                    )
-                } else {
-                    visitMethodInsn(opcode, owner, name, desc, itf)
-                }
-            }
-    }
-
-    /**
-     * Replaces `System.nanoTime` and `System.currentTimeMillis` with stubs to prevent non-determinism
-     */
-    private class DeterministicTimeTransformer(val adapter: GeneratorAdapter) : MethodVisitor(ASM_API, adapter) {
-        override fun visitMethodInsn(opcode: Int, owner: String, name: String, desc: String, itf: Boolean) =
-            adapter.run {
-                if (owner == "java/lang/System" && (name == "nanoTime" || name == "currentTimeMillis")) {
-                    invokeIfInTestingCode(
-                        original = { visitMethodInsn(opcode, owner, name, desc, itf) },
-                        code = { push(1337L) } // any constant value
-                    )
-                    return
-                }
-                visitMethodInsn(opcode, owner, name, desc, itf)
-            }
-    }
-
-    /**
-     * Makes java.util.Random and all classes that extend it deterministic.
-     * In every Random method invocation replaces the owner with Random from ManagedStateHolder.
-     */
-    private inner class DeterministicRandomTransformer(
-        fileName: String,
-        className: String,
-        methodName: String,
-        adapter: GeneratorAdapter,
-    ) :
-        ManagedStrategyMethodVisitor(fileName, className, methodName, adapter) {
-        override fun visitMethodInsn(opcode: Int, owner: String, name: String, desc: String, itf: Boolean) =
-            adapter.run {
-                if (owner == "java/util/concurrent/ThreadLocalRandom" ||
-                    owner == "java/util/concurrent/atomic/Striped64" ||
-                    owner == "java/util/concurrent/atomic/LongAdder" ||
-                    owner == "java/util/concurrent/atomic/DoubleAdder" ||
-                    owner == "java/util/concurrent/atomic/LongAccumulator" ||
-                    owner == "java/util/concurrent/atomic/DoubleAccumulator"
-                ) {
-                    if (name == "nextSecondarySeed" || name == "getProbe") { // INVOKESTATIC
-                        invokeIfInTestingCode(
-                            original = {
-                                visitMethodInsn(opcode, owner, name, desc, itf)
-                            },
-                            code = {
-                                invokeStatic(Injections::nextInt)
-                            }
-                        )
-                        return
-                    }
-                    if (name == "advanceProbe") { // INVOKEVIRTUAL
-                        invokeIfInTestingCode(
-                            original = {
-                                visitMethodInsn(opcode, owner, name, desc, itf)
-                            },
-                            code = {
-                                pop()
-                                invokeStatic(Injections::nextInt)
-                            }
-                        )
-                        return
-                    }
-                    if (name == "nextInt" && desc == "(II)I") {
-                        invokeIfInTestingCode(
-                            original = {
-                                visitMethodInsn(opcode, owner, name, desc, itf)
-                            },
-                            code = {
-                                val arguments = storeArguments(desc)
-                                pop()
-                                loadLocals(arguments)
-                                invokeStatic(Injections::nextInt2)
-                            }
-                        )
-                        return
-                    }
-                }
-                if (isRandomMethod(name, desc)) {
-                    invokeIfInTestingCode(
-                        original = {
-                            visitMethodInsn(opcode, owner, name, desc, itf)
-                        },
-                        code = {
-                            val arguments = storeArguments(desc)
-                            val ownerLocal = newLocal(getType("L$owner;"))
-                            storeLocal(ownerLocal)
-                            ifStatement(
-                                condition = {
-                                    loadLocal(ownerLocal)
-                                    invokeStatic(Injections::isRandom)
-                                },
-                                ifClause = {
-                                    invokeInIgnoredSection {
-                                        invokeStatic(Injections::deterministicRandom)
-                                        loadLocals(arguments)
-                                        /*
-                                        In Java 21 RandomGenerator interface was introduced so sometimes data structures
-                                        interact with java.util.Random through this interface.
-                                         */
-                                        val randomOwner = if (owner.endsWith("RandomGenerator")) "java/util/random/RandomGenerator" else "java/util/Random"
-                                        visitMethodInsn(opcode, randomOwner, name, desc, itf)
-                                    }
-                                },
-                                elseClause = {
-                                    loadLocal(ownerLocal)
-                                    loadLocals(arguments)
-                                    visitMethodInsn(opcode, owner, name, desc, itf)
-                                }
-                            )
-                        }
-                    )
-                    return
-                }
-                visitMethodInsn(opcode, owner, name, desc, itf)
-            }
-
-        private fun isRandomMethod(methodName: String, desc: String): Boolean = randomMethods.any {
-            it.name == methodName && it.descriptor == desc
-        }
-    }
-
-    private companion object {
-        private val randomMethods =
-            Random::class.java.declaredMethods.map { Method.getMethod(it) }
-    }
-
-    /**
      * Adds invocations of ManagedStrategy methods before park and after unpark calls
      */
     private inner class ParkUnparkTransformer(
